requirements.txt
nxt-python
#PyBluez
